Implemented CPU
CPU組み込み完了、ただし、msx_slot のタイミングとズレているので動かず

+++ controller/FPGA_MSXtR_Stack_Controller_001/fpga_io.h
@@ -52,6 +52,6 @@ void fpga_msx_pause( bool pause_on );
 void fpga_bootrom_enable( bool enable );
 void fpga_set_bus_owner( uint8_t owner );
 void fpga_set_keyboard_matrix( const uint8_t *matrix );
-uint8_t fpga_get_debug_signal( void );
+uint16_t fpga_get_debug_signal( void );
 
 #endif
